adjacent-app/requirements.txt
fastapi
uvicorn
pydantic
pandas
python-dotenv
sentence-transformers
torch
faiss-cpu
networkx
pytest
